+++ api/src/main/java/org/allaymc/api/entity/interfaces/EntityAreaEffectCloud.java
@@ -1,7 +1,9 @@
 package org.allaymc.api.entity.interfaces;
 
 import org.allaymc.api.entity.Entity;
+import org.allaymc.api.entity.component.EntityAreaEffectCloudBaseComponent;
+import org.allaymc.api.entity.component.EntityPotionComponent;
 
-public interface EntityAreaEffectCloud extends Entity {
+public interface EntityAreaEffectCloud extends Entity, EntityAreaEffectCloudBaseComponent, EntityPotionComponent {
 
 }