+++ api/src/main/java/org/allaymc/api/entity/interfaces/EntityArrow.java
@@ -1,6 +1,7 @@
 package org.allaymc.api.entity.interfaces;
 
 import org.allaymc.api.entity.component.EntityArrowBaseComponent;
+import org.allaymc.api.entity.component.EntityPotionComponent;
 
-public interface EntityArrow extends EntityLiving, EntityProjectile, EntityArrowBaseComponent {
+public interface EntityArrow extends EntityLiving, EntityProjectile, EntityArrowBaseComponent, EntityPotionComponent {
 }