+++ api/src/main/java/org/allaymc/api/entity/interfaces/EntityLingeringPotion.java
@@ -1,7 +1,7 @@
 package org.allaymc.api.entity.interfaces;
 
-import org.allaymc.api.entity.Entity;
+import org.allaymc.api.entity.component.EntityPotionComponent;
 
-public interface EntityLingeringPotion extends Entity {
+public interface EntityLingeringPotion extends EntityProjectile, EntityPotionComponent {
 
 }