+++ api/src/main/java/org/allaymc/api/item/interfaces/ItemLingeringPotionStack.java
@@ -1,6 +1,8 @@
 package org.allaymc.api.item.interfaces;
 
 import org.allaymc.api.item.ItemStack;
+import org.allaymc.api.item.component.ItemPotionComponent;
+import org.allaymc.api.item.component.ItemProjectileComponent;
 
-public interface ItemLingeringPotionStack extends ItemStack {
+public interface ItemLingeringPotionStack extends ItemStack, ItemProjectileComponent, ItemPotionComponent {
 }